==> research/1_Get_GPS_Cords.ipynb <==
import pandas as pd
from geopy.geocoders import Nominatim

df = pd.read_csv('../data/data.csv')
df.head()

# calling the Nominatim tool
loc = Nominatim(user_agent="GetLoc")

location = df['ADDRESS'].to_list()
location[0]

# entering the location name
getLoc = loc.geocode(location[0])

print(getLoc)

getLoc.latitude

getLoc.longitude

for town in location:
    getLoc = loc.geocode(town)
    if getLoc is not None:
        df.loc[df['ADDRESS'] == town, 'LATITUDE'] = getLoc.latitude
        df.loc[df['ADDRESS'] == town, 'LONGITUDE'] = getLoc.longitude
        print(f"ADDRESS {town}: cords {getLoc} {getLoc[1]}")
    else:
        df.loc[df["ADDRESS"] == town, "LATITUDE"] = None
        df.loc[df["ADDRESS"] == town, "LONGITUDE"] = None

df.head()

df.to_csv('../data/data_gps.csv', index=False)

# calling the nominatim tool
geoLoc = Nominatim(user_agent="GetLoc")

# passing the coordinates
locname = geoLoc.reverse("7.4307776, 79.9298775")

# printing the address/location name
print(locname.address)
