==> research/3_Create_Matrix_json.ipynb <==
import pandas as pd
import requests
import numpy as np
import time
import folium

df = pd.read_csv('../data/data_gps.csv')

df.columns

if 'ADDRESS' in df.columns and 'LATITUDE' in df.columns and 'LONGITUDE' in df.columns:
    unique_locations_df = df.drop_duplicates(subset=["ADDRESS"])[
        ["ADDRESS", "LATITUDE", "LONGITUDE"]
    ]
    
else:
    print("Address, Latitude, Longitude columns not found in the data.")
    unique_locations_df = None

unique_locations_df.shape

len(unique_locations_df)

kiribathkumbura = (7.2726897, 80.5618714)
kiribathkumbura_data = {
    "ADDRESS": "Kiribathkumbura",
    "LATITUDE": kiribathkumbura[0],
    "LONGITUDE": kiribathkumbura[1],
}

unique_locations_df = pd.concat(
    [unique_locations_df, pd.DataFrame([kiribathkumbura_data])], ignore_index=True
)
unique_locations_df.shape

def get_osrm_data(origin, destination):
    """
    Get the distance between two coordinates using OSRM API.
    :param origin: (latitude, longitude)
    :param destination: (latitude, longitude)
    :return: Distance in meters
    """
    osrm_base_url = "http://router.project-osrm.org/route/v1/car"
    url = f"{osrm_base_url}/{origin[1]},{origin[0]};{destination[1]},{destination[0]}?overview=full&geometries=geojson"
    responce = requests.get(url)

    if responce.status_code == 200:
        data = responce.json()
        if "routes" in data and len(data["routes"]) > 0:
            path_cords = data["routes"][0]["geometry"]["coordinates"]
            distance = data["routes"][0]["distance"] / 1000
            return path_cords, distance
        
    return None, np.inf

location_names = unique_locations_df['ADDRESS'].to_list()
location_cords = list(zip(unique_locations_df["LATITUDE"], unique_locations_df["LONGITUDE"]))

origin = location_cords[0]
destination = location_cords[1]
get_osrm_data(origin, destination)

loc_count = len(location_cords)
loc_count

distance_matrix = np.zeros((loc_count, loc_count))
route_data = []

for i in range(loc_count):
    for j in range(i+1, loc_count):
        origin = location_cords[i]
        destination = location_cords[j]

        path_cords, distance = get_osrm_data(origin, destination)

        distance_matrix[i][j] = distance
        distance_matrix[j][i] = distance

        route_data.append(
            {
                "origin": location_names[i],
                "destination": location_names[j],
                "distance": distance,
                "path_cords": path_cords,
            }
        )
        
        time.sleep(0.2)

import json

# Prepare the data for JSON storage
distance_matrix_data = {
    "locations": location_names,  # Store location names
    "cords": location_cords,
    "route_data": route_data,  # Store route data as a list of dictionaries
    "distance_matrix": distance_matrix.tolist(),  # Convert NumPy array to list for JSON
}

# Define JSON file path
json_file_path = "../data/osrm_distance_matrix.json"

# Save to JSON file
with open(json_file_path, "w") as json_file:
    json.dump(distance_matrix_data, json_file, indent=4)

# Display the JSON file path
json_file_path
