==> research/4_Load_json.ipynb <==
import json

file_path = '../data/osrm_distance_matrix.json'

# Open a JSON file and load its contents into a Python dictionary
with open(file_path, "r") as file:
    data = json.load(file)

data['distance_matrix']

origin = "Medawachchiya"
destination =  "Nochchiyagama"

route_data = data['route_data']

for route in route_data:
    if route['origin'] == origin and route['destination'] == destination:
        print(route)
        break
    elif route["origin"] == destination and route["destination"] == origin:
        print(route)
        break

def load_json(file_path):
    with open(file_path, 'r') as file:
        return json.load(file)


def find_route(route_data, origin, destination):
    # data = load_json(file_path)
    # route_data = data['route_data']
    for route in route_data:
        if route['origin'] == origin and route['destination'] == destination:
            return route
        elif route["origin"] == destination and route["destination"] == origin:
            return route
    return None

data = load_json(file_path=file_path)

distance_matrix = data['distance_matrix']

distance_matrix
